Added keyvalue store and fixed unit tests
Added keyvalue store and fixed unit tests

this involved redesigning the CustomerService classes, and the corresponding changes to dependent services.

+++ src/main/java/com/btb/coffeeshop/purchase/InsufficientFundsException.java
@@ -7,11 +7,11 @@ public class InsufficientFundsException extends Exception {
 	 */
 	private static final long serialVersionUID = 59631943927460796L;
 
-	private Long customerId;
+	private Integer customerId;
 	private Integer orderTotalPrice;
 	private Integer customerBalance;
 
-	public InsufficientFundsException(Long id, Integer totalPrice, Integer balance) {
+	public InsufficientFundsException(Integer id, Integer totalPrice, Integer balance) {
 		this.customerBalance = balance;
 		this.customerId = id;
 		this.orderTotalPrice = totalPrice;
